colorpalette/requirements.txt
Bootstrap_Flask==2.2.0
Flask_WTF==1.2.1
WTForms==3.0.1
Werkzeug==3.0.0
Flask==2.3.2
pillow==10.2.0
numpy==1.26.4
webcolors==1.13
scipy~=1.12.0